Full Stack Open Part 2a: Exercises 2.5: Declare Course component as its own module and refactor Course and all sub-components into Course.jsx. Import Course into App component
